fix: use MINGW instead of MSVC in CMakeLists.txt
chore: include line strats in nomake.bat

fix: define custom defined pi/4

refactor: make smoothing factor as double floating point

+++ src/artist/CardinalSplineArtist.h
@@ -5,7 +5,7 @@
 class CardinalSplineArtist : public Artist {
 private:
     std::vector<utils::Point<int>> control_points;
-    int c = 1;
+    double c = 1;
 
     void drawCardinalSpline(HDC hdc, const std::vector<utils::Point<int>>& pts, double c ,COLORREF color);
